refactor : rename enum class and update it to use for constant enum
